pyproject.toml
[project]
name = "llm-as-labelers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "matplotlib>=3.10.5",
    "onnx>=1.18.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
